- 新增React相關套件 - 分類 frontend 中的 components - 將 scripts 更名為 pages - 移除navbar測試檔 - 更改首頁路徑 - 完成登入狀態的下拉選單
待完成

- 登入的下拉選單點擊無法去指定頁面
- 登入/註冊  popupwindow
