--- app/frontend/components/MenuItems.js
@@ -1,15 +1,0 @@
-export const MenuItems = [
-  {
-    title:'登入',
-    path:'',
-    cName :'dropdown-link'
-
-  },
-  {
-    title:'註冊',
-    path:'',
-    cName :'dropdown-link'
-
-  }
-
- ]
