+++ app/frontend/components/navbar/Button.scss
@@ -43,12 +43,10 @@
    
     font-size: 18px;
     color: #fff;
-
     
   }
 
 }
-
 
 
 @media screen and (max-width: 960px) {
