+++ app/frontend/components/navbar/Dropdown.scss
