+++ app/frontend/components/navbar/Navbar.css
