+++ app/frontend/components/navbar/Navbar.js
@@ -1,0 +1,68 @@
+import React,{useState} from 'react';
+import { Button } from './Button';
+import {Link} from 'react-router-dom'
+import  './Navbar.css'
+
+
+
+
+
+const Navbar = ({user}) => {
+
+
+  const [click,setClick] = useState(false)
+  const [dropdown,setDropdown] = useState(false)
+
+  const handleClick = () => { setClick(!click)}
+  const closeMobileMenu = () => {setClick(false)};
+
+  
+  return(
+    <>
+    <nav className="navbar">
+    <Link to='/' className="navbar-logo">
+      CallBack
+    </Link>
+
+    <div className="menu-icon" onClick={handleClick}>
+
+      <i className={click ? 'fas fa-times':'fas fa-bars'}></i>
+
+    </div>
+
+    <ul className={click? 'nav-menu active':'nav-menu'}>
+
+      <li className="nav-items  nav-links" onClick={() => {window.location.href="/about"}}>關於</li>
+      <li className="nav-items  nav-links" onClick={() => {window.location.href="/project/proposal"}}>提案</li>
+      <li className="nav-items nav-links" onClick={() => {window.location.href="/projects"}}>探索</li>
+
+      <li className="nav-items">
+        <Link to='/' className='nav-links-mobile' onClick={closeMobileMenu}>註冊</Link>
+      </li>
+      
+    </ul>
+    <Button user={user}/>
+   
+    </nav>
+    </>
+  )
+  
+}
+
+export default Navbar;
+
+// <li className="nav-items  nav-links" onClick={() => {window.location.href="/about"}}>關於</li>
+// <li className="nav-items  nav-links" onClick={() => {window.location.href="/project/proposal"}}>提案</li>
+// <li className="nav-items  nav-links" onClick={() => {window.location.href="/projects"}}>探索</li>
+
+// <li className="nav-items  nav-links" onClick={closeMobileMenu}>
+// <Link to='/' className='nav-links' onClick={closeMobileMenu}>關於</Link>
+// </li>
+
+// <li className="nav-items">
+//   <Link to='/' className='nav-links' onClick={closeMobileMenu}>提案</Link>
+// </li>
+
+// <li className="nav-items">
+//   <Link to='/' className='nav-links' onClick={closeMobileMenu}>探索</Link>
+// </li>
