+++ app/frontend/pages/index.js
@@ -1,0 +1,2 @@
+import './home'
+import './navbar'
