+++ app/frontend/pages/navbar/index.js
@@ -6,9 +6,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types'
 
-
-import App from '../scripts/App'
-import  '../styles/application.scss'
+import App from '../App'
+import '../../styles/application.scss'
 
 
 
@@ -17,8 +16,8 @@ document.addEventListener('turbolinks:load', () => {
    
   ReactDOM.render(
     <App/>,
-    document.getElementById('root'),
-  
+    document.getElementById('navbar_container'),
   )
   
 })
+
