+++ app/frontend/styles/application.scss
@@ -9,3 +9,4 @@
 
 
 @import "./navbar";
+
