+++ app/frontend/styles/navbar.scss
@@ -20,7 +20,6 @@ nav{
     .nav-left{
 
 
-
       .menu-item{
         
         display: inline-block;
