+++ app/views/shared/_navbar.html
@@ -1,0 +1,4 @@
+
+<div id="navbar_container"></div>
+
+
