--- app/views/shared/_navbar.html.erb
@@ -1,3 +1,0 @@
-
-<div id="root"></div>
-
