+++ config/routes.rb
@@ -1,5 +1,5 @@
 Rails.application.routes.draw do
-  root to: "projects#index"
+  root to: "home#index"
 
   devise_for :users, controllers: { 
     sessions: 'users/sessions',
