+++ package.json
@@ -3,7 +3,8 @@
   "private": true,
   "dependencies": {
     "@babel/preset-react": "^7.12.10",
-    "@craco/craco": "^6.0.0",
+    "@emotion/react": "^11.1.4",
+    "@emotion/styled": "^11.0.0",
     "@fortawesome/fontawesome-free": "^5.15.1",
     "@fullhuman/postcss-purgecss": "^3.1.3",
     "@rails/actioncable": "^6.0.0",
@@ -11,15 +12,17 @@
     "@rails/ujs": "^6.0.0",
     "@rails/webpacker": "4.3.0",
     "@tailwindcss/postcss7-compat": "^2.0.2",
+    "antd": "^4.9.4",
     "autoprefixer": "^9.8.6",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "framer-motion": "^3.1.1",
     "postcss": "^7.0.35",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-elastic-carousel": "^0.10.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.1",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.2",
+    "styled-components": "^5.2.1",
     "turbolinks": "^5.2.0"
   },
   "version": "0.1.0",
